+++ CHANGELOG.md
@@ -20,8 +20,15 @@ Please follow <https://changelog.md/> conventions.
 
 ## 0.2.0
 
-- Update CMTAT to [v3.0.0-rc2](https://github.com/CMTA/CMTAT/releases/tag/v3.0.0-rc2)
-- Update OpenZeppelin library to [v5.3.0](https://github.com/OpenZeppelin/openzeppelin-contracts/releases/tag/v5.3.0)
+- Dependencies
+  - Update CMTAT to [v3.0.0-rc7](https://github.com/CMTA/CMTAT/releases/tag/v3.0.0-rc2)
+  - Update OpenZeppelin library to [v5.4.0](https://github.com/OpenZeppelin/openzeppelin-contracts/releases/tag/v5.4.0)
+  - Update Solidity to 0.8.30
+  - Update EVM version to Prague
+
+- Better code separation trough the creation of modules
+
+- Use [ERC-7201](https://eips.ethereum.org/EIPS/eip-7201) for storage for compatibility with CMTAT
 
 ## 0.1.0